+++ src/main/java/mips/DataMemory.java
@@ -35,10 +35,7 @@ public class DataMemory implements MemoryUnit {
   @Override
   public int getMemoryUnit(int index) {
     if (index % this.wordSize == 0) {
-      return this.memory[
-          index
-              / this.wordSize]; // This is always an integer since result of mod is 0, don't need to
-      // cast or floor
+      return this.memory[index / this.wordSize];
     } else {
       throw new IllegalArgumentException(
           String.format(
@@ -50,7 +47,7 @@ public class DataMemory implements MemoryUnit {
   @Override
   public void setMemoryUnit(int index, int value) {
     if (index % this.wordSize == 0) {
-      this.memory[(int) (index / this.wordSize)] = value;
+      this.memory[index / this.wordSize] = value;
     } else {
       throw new IllegalArgumentException(
           String.format(
